+++ MemoryInspector/MemoryInspectorFPP8AEPFloatingPoint.m
@@ -122,12 +122,11 @@
 	unsigned long long mantissa = [f mantissa];
 	*value = [NSArray arrayWithObjects:
 		[NSNumber numberWithInt:077770000 | [f exponent]],
-		[NSNumber numberWithInt:
-			077770000 | ([f negative] ? 04000 : 0) | (unsigned long) (mantissa >> 53)],
-		[NSNumber numberWithInt:077770000 | (unsigned long) ((mantissa >> 41) & 07777)],
-		[NSNumber numberWithInt:077770000 | (unsigned long) ((mantissa >> 29) & 07777)],
-		[NSNumber numberWithInt:077770000 | (unsigned long) ((mantissa >> 17) & 07777)],
-		[NSNumber numberWithInt:077770000 | (unsigned long) ((mantissa >> 5) & 07777)],
+		[NSNumber numberWithInt:(int)(077770000 | ([f negative] ? 04000 : 0) | (unsigned long) (mantissa >> 53))],
+		[NSNumber numberWithInt:(int)(077770000 | (unsigned long) ((mantissa >> 41) & 07777))],
+		[NSNumber numberWithInt:(int)(077770000 | (unsigned long) ((mantissa >> 29) & 07777))],
+		[NSNumber numberWithInt:(int)(077770000 | (unsigned long) ((mantissa >> 17) & 07777))],
+		[NSNumber numberWithInt:(int)(077770000 | (unsigned long) ((mantissa >> 5) & 07777))],
 		nil];
 	return TRUE;
 }
